Fix QDateEditEx's minimumSizeHint so the clear button doesn't hide the year
Adjusted the test program to show the issue.

Fixed the positioning of the clear button to work better with styles that
have some space on the right of the "up" button, the old calculation assumed
there wasn't any, so the clear button could end up too much on the right,
almost over the arrow-down button.

Sent upstream at https://github.com/doberkofler/QDateEditEx


git-svn-id: svn+ssh://svn.kdab.com/home/SVN-klaralv/projects/FatCRM/trunk@62135 c66179b2-ff0d-0410-8aef-1b1a91e65cfa