Major cleanup: share keys in KDCRMFields, streamline Account serialization
As I previously did in Note and Email: move the accessor hash for accounts
to the shared lib kdcrmdata, and use it for setData/data and accountio.cpp.
This required changing the name of the widgets (e.g. from accountType
to account_type).

This reduces code duplication and allowed to add support for custom
fields in accounts more easily.