FATCRM-64: Rework syncing algorithm to handle remote-deletion.
Because RID REMOVE doesn't work correctly in akonadiserver, we need
to resolve IDs on deleted items. This requires an ItemFetchJob, which
can't run within the ItemSync. So fetching deleted items is now done
in a custom task, which then resolves the items and finally runs
the ItemDeleteJob.

Also splitted the last-updated timestamp from the last-deleted
timestamp to avoid fetching the same deleted items over and over again
(leading to "item not found" errors).

For the same reason, also add 1s to the timestamp for the next time,
to avoid getting the same items again.