FATCRM-74: fix mixup when creating accounts or contacts.
The app first gets notified with an empty ID, then the resource gets the ID
from the server and the app gets a dataChanged. So we can't do anything
that requires the ID until the dataChanged comes in.

This fixes the duplicate entry in the account combo after creating an account
(the first item being basically unusable). Same issue with "reports to" for contacts.