Get rid of AccountCountryRef and use AccountRepository instead.
This is a more scalable solution for adding other fields coming from
the account into the opp list (FATCRM-95).

It also adds signals to AccountRepository which will be useful for
detecting that the account now has been created by the resource
(to get its remote ID) in the csv importer.