Fix previous commit so it also works with existing resources.
The new loading overlay shows that modelLoaded() wasn't correctly emitted
in all cases.

For a new resource, we now get the "populated" signal right away but ignore it
because no rows, and we consider the first chunk of contacts as
"ok contacts are populated now" (not fully correct, but better than nothing).

For an existing resource, we wait until isCollectionPopulated returns true.

FATCRM-150