Port the resource to version 4.1 of the Sugar SOAP interface.
This allows to get account_id in opportunities, which made it possible
to remove all the hacks around resolving account names.

It also removes the need for ListDeletedEntriesJob, since the normal
listing (with deleted=true) now includes both normal and deleted items.

This commit also includes the fix for handling the merging of accounts
(which doesn't update the associated opportunities)