Remove one roundtrip when listing entries; remove +1s hack.
Calling incrementTimestamp is incorrect, as detected by
test_jobwithsugarsoapprotocol: in the same second, we can
modify account A, list accounts, and then modify account A.
If we do that, then any later listing would then not see the second
modification, because of "lastModified >= [1s later]".

So we have to get the same last-modified object every time.
We could optimize away sending it to akonadi (by comparing with the last
object, using SugarAccount::operator==), but that's it, we can't do any
better.

The ideal solution would be a SQL filter like
"account.date_modified >= timestamp AND account.etag != lastetag"
but that isn't available (and even with REST ETAGs, that would likely
not be available at the DB query level anyhow).