models: See which opps are 'warm'
Indicate whether we are past a next step date or whether we recently
updated the last modified date.

Try to do that in a performant way; calculating the background color on
the fly won't work since we'd have to ask for the current date time for
each row individual in that case. Thus create a cache in ItemsTreeModel where
we store the backgrounds for all rows.

Task-Id: FATCRM-109