opps: Allow filter "Closed Won" and "Closed Lost"
We have a use-case where we only would like to look at closed lost or
closed won opps. Make it easy to filter by them via the filter widget.

It probably makes less sense for the other less used sales stages. In
theory the right widget would be a drop-down list of check boxes with all
individual sales stages, but that's probably overkill.