Port to QtKeychain
Currently we optionally use KWallet to store passwords and fall back to
storing it in a settings file.

Instead use QtKeychain to store the password in the platforms respective
native API (which happens to be KWallet on Plasma).

This should improve the cross-platformness of the app and matches the
general direction KDE is taking wrt KWallet.

It also considerably simplifies the implifies the implementation since
we no longer need to handle the optionalness of KWallet